Update ht_sols_microscope_gui.py
Added a GUI for the XY stage with buttons to move up/down and left/right as a user defined percentage of the current FOV (which enables precise motion). Also included are 'last move direction' and 'current position' text boxes to aid navigation. The 'scout mode' has been re-worked to enable smooth navigation with both the hardware enabled joystick and the new GUI. An additional argument was added to 'snap_volume' to reduce lines of code.