Update ht_sols_microscope_gui.py
Bug fix -> previously laser settings were only applied to the microscope if the 'power_per_channel' variable was changed. So it was possible (with lasers of the same power) to toggle between them with the 'checkbox' and the microscope settings would NOT update. To fix this there is now an explicit check on the status of the checkbox using the 'channels_per_slice' variable.