Update ht_sols_microscope.py
Setting 'fast_init=False' for the zoom lens since one of the controllers is flakey and doesn't always start properly. The slow init should be more robust to power cycle and software restarts. The zoom lens is typically reliable once homed, and so this setting should not affect the rest of the imaging session (since the home routine only runs after a power cycle).