Update ht_sols_microscope_gui.py
Hovertips for 'tkinter_compound_widgets' (tkcw) are now handled within tkcw to avoid conflict between the <leave> binding that is nested in tkcw to update fields. Otherwise there is a risk that a hovertip binding matches geometrically with the <leave> binding (e.g. they target the same frame) and then either one or the other is not triggered.